// Multitier/Lernplattform/web/src/app/api/courses/model/courseOverview.ts
/**
 * Course Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CourseOverview { 
    id?: string;
    name?: string;
    deleted?: boolean;
    schema?: number;
}